# gym-track-angular/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification",
    "description": "OpenApi documentation",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/plan-exercise": {
      "post": {
        "tags": [
          "plan-exercise-controller"
        ],
        "operationId": "savePlanExercise",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanExerciseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/plan-exercise/workout/{id}": {
      "post": {
        "tags": [
          "plan-exercise-controller"
        ],
        "operationId": "savePlanExerciseForSpecificWorkout",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanExerciseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/notes": {
      "get": {
        "tags": [
          "notes-controller"
        ],
        "operationId": "findAllNotes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotesResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "notes-controller"
        ],
        "operationId": "saveNotes",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/exercise": {
      "post": {
        "tags": [
          "exercise-controller"
        ],
        "operationId": "saveExercise",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
...
    "/workout-logs/exercise/{id}": {
      "get": {
        "tags": [
          "workout-logs-controller"
        ],
        "operationId": "findAllLogsByExerciseIdAndUserId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkoutLogsResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/plan/{plan_id}": {
      "get": {
        "tags": [
          "plan-controller"
        ],
        "operationId": "findPlanById",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/plan-exercise/{plan-id}": {
      "get": {
        "tags": [
          "plan-exercise-controller"
        ],
        "operationId": "findAllPlanExercisesByPlanId",
        "parameters": [
          {
            "name": "plan-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanExerciseResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/notes/{id}": {
      "get": {
        "tags": [
          "notes-controller"
        ],
        "operationId": "findAllNotesForExercise",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotesResponse"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "notes-controller"
        ],
        "operationId": "deleteNotes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/exercise/{exercise_id}": {
      "get": {
        "tags": [
          "exercise-controller"
        ],
        "operationId": "findExerciseById",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/exercise/type/{type}": {
      "get": {
        "tags": [
          "exercise-controller"
        ],
        "operationId": "findAllExercisesByType",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "CHEST",
                "BACK",
                "BICEPS",
                "SHOULDERS",
                "TRICEPS",
                "FOREARM",
                "ABS",
                "GLUTES",
                "CALF",
                "THIGH",
                "HAMSTRINGS"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExerciseResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/exercise/my-exercises/{type}": {
      "get": {
        "tags": [
          "exercise-controller"
        ],
        "operationId": "findAllUserExercisesByType",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "CHEST",
                "BACK",
                "BICEPS",
                "SHOULDERS",
                "TRICEPS",
                "FOREARM",
                "ABS",
                "GLUTES",
                "CALF",
                "THIGH",
                "HAMSTRINGS"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExerciseResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/plan-exercise/{id}": {
      "delete": {
        "tags": [
          "plan-exercise-controller"
        ],
        "operationId": "deletePlanExercise",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/exercise/{id}": {
      "delete": {
        "tags": [
          "exercise-controller"
        ],
        "operationId": "deletePlanExercise_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PlanExerciseRequest": {
        "required": [
          "exerciseId"
        ],
        "type": "object",
        "properties": {
          "sets": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "exerciseId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "NotesRequest": {
        "required": [
          "exerciseId",
          "note"
        ],
        "type": "object",
        "properties": {
          "note": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "exerciseId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ExerciseRequest": {
        "required": [
          "exerciseType",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "exerciseType": {
            "type": "string",
            "enum": [
              "CHEST",
              "BACK",
              "BICEPS",
              "SHOULDERS",
              "TRICEPS",
              "FOREARM",
              "ABS",
              "GLUTES",
              "CALF",
              "THIGH",
              "HAMSTRINGS"
            ]
          }
        }
      },
...
      "WorkoutLogsResponse": {
        "type": "object",
        "properties": {
          "setNumber": {
            "type": "integer",
            "format": "int32"
          },
          "weight": {
            "type": "number",
            "format": "double"
          },
          "reps": {
            "type": "integer",
            "format": "int32"
          },
          "summaryWeight": {
            "type": "number",
            "format": "double"
          },
          "exerciseId": {
            "type": "integer",
            "format": "int64"
          },
          "sessionId": {
            "type": "integer",
            "format": "int64"
          },
          "sessionDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "PlanResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "exercises": {
            "type": "integer",
            "format": "int32"
          },
          "sets": {
            "type": "integer",
            "format": "int32"
          },
          "muscleGroups": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "CHEST",
                "BACK",
                "BICEPS",
                "SHOULDERS",
                "TRICEPS",
                "FOREARM",
                "ABS",
                "GLUTES",
                "CALF",
                "THIGH",
                "HAMSTRINGS"
              ]
            }
          }
        }
      },
      "PlanExerciseResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "exerciseName": {
            "type": "string"
          },
          "exerciseId": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string",
            "enum": [
              "CHEST",
              "BACK",
              "BICEPS",
              "SHOULDERS",
              "TRICEPS",
              "FOREARM",
              "ABS",
              "GLUTES",
              "CALF",
              "THIGH",
              "HAMSTRINGS"
            ]
          },
          "sets": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ExerciseResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "CHEST",
              "BACK",
              "BICEPS",
              "SHOULDERS",
              "TRICEPS",
              "FOREARM",
              "ABS",
              "GLUTES",
              "CALF",
              "THIGH",
              "HAMSTRINGS"
            ]
          }
        }
      },
      "NotesResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "note": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}